requirements.txt
django
requests
python-dotenv
pytest
pytest-django
web3
py
titanoboa

